functions/download/requirements.txt
psycopg2-binary
yt-dlp
ffmpeg
boto3
requests